// src/types/predictions.ts
export interface PredictionData {
  id: number;
  household_id: string;
  region: string;
  district: string;
  cluster: string;
  village: string;
  cohort: string;
  cycle: string;
  latitude: number;
  longitude: number;
  altitude: number;
  standard_evaluation: boolean;
  checkin_evaluation: boolean;
  evaluation_month: number;
  Land_size_for_Crop_Agriculture_Acres: number;
  farm_implements_owned: number;
  tot_hhmembers: number;
  Distance_travelled_one_way_OPD_treatment: number;
  Average_Water_Consumed_Per_Day: number;
  hh_water_collection_Minutes: number;
  vsla_participation: number;
  ground_nuts: number;
  composts_num: number;
  perennial_crops_grown_food_banana: number;
  sweet_potatoes: number;
  perennial_crops_grown_coffee: number;
  irish_potatoes: number;
  business_participation: number;
  cassava: number;
  hh_produce_lq_manure: number;
  hh_produce_organics: number;
  maize: number;
  sorghum: number;
  non_bio_waste_mgt_present: number;
  soap_ash_present: number;
  education_level_encoded: number;
  tippy_tap_present: number;
  hhh_sex: number;
  prediction: number;
  predicted_income: number;
  probability: number;
  date: string;
}